added temporary webpage state management solution
